[expo][refactor] move context implementation to it's feature directory
**Summary**

we decided a minor structure change. all context implementation is moved to their feature directory.

`features/xxxx/context` provide a context functions for the feature xxxx.

**Test**

- expo
- yarn test

**Issue**

- N/A
