+++ expo/features/artist/ByGroupView.tsx
@@ -1,5 +1,5 @@
-import { useHelloProject } from '@hpapp/contexts/serviceroot';
 import ArtistCard from '@hpapp/features/artist/ArtistCard';
+import { useHelloProject } from '@hpapp/features/root/protected/context';
 import { FlatList } from 'react-native';
 
 export default function ByGroupView() {
