+++ expo/features/artist/sort/HPSortResultListItem.tsx
@@ -1,9 +1,9 @@
-import { useHelloProject } from '@hpapp/contexts/serviceroot';
 import FollowIcon from '@hpapp/features/artist/FollowIcon';
 import MemberIcon from '@hpapp/features/artist/MemberIcon';
 import HPSortResultRankDiffIcon from '@hpapp/features/artist/sort/HPSortResultRankDiffIcon';
 import Text from '@hpapp/features/common/components/Text';
 import { Spacing, FontSize, MemberIconSize, IconSize } from '@hpapp/features/common/constants';
+import { useHelloProject } from '@hpapp/features/root/protected/context';
 import { Divider } from '@rneui/themed';
 import { StyleSheet, View } from 'react-native';
