+++ expo/features/artist/sort/HPSortResultListTopItem.tsx
@@ -1,9 +1,9 @@
-import { useHelloProject } from '@hpapp/contexts/serviceroot';
 import FollowIcon from '@hpapp/features/artist/FollowIcon';
 import MemberIcon from '@hpapp/features/artist/MemberIcon';
 import HPSortResultRankDiffIcon from '@hpapp/features/artist/sort/HPSortResultRankDiffIcon';
 import Text from '@hpapp/features/common/components/Text';
 import { FontSize, IconSize, MemberIconSize, Spacing } from '@hpapp/features/common/constants';
+import { useHelloProject } from '@hpapp/features/root/protected/context';
 import { Icon } from '@rneui/themed';
 import { View, StyleSheet } from 'react-native';
