+++ expo/features/root/TestRoot.tsx
@@ -1,8 +1,8 @@
-import { AnalyticsProvider } from '@hpapp/contexts/analytics';
-import { RelayProvider } from '@hpapp/contexts/relay';
-import { SettingsProvider } from '@hpapp/contexts/settings';
-import { LocalUserConfigurationSettings } from '@hpapp/contexts/settings/useLocalUserConfig';
 import { CurrentUserSettings } from '@hpapp/features/auth';
+import { AnalyticsProvider } from '@hpapp/features/root/context/analytics';
+import { RelayProvider } from '@hpapp/features/root/context/relay';
+import { SettingsProvider } from '@hpapp/features/settings/context';
+import { LocalUserConfigurationSettings } from '@hpapp/features/settings/context/useLocalUserConfig';
 
 const settings = [CurrentUserSettings, LocalUserConfigurationSettings];
