+++ expo/features/root/protected/ProtectedRoot.tsx
@@ -1,10 +1,10 @@
-import { ServiceRootProvider } from '@hpapp/contexts/serviceroot';
-import { AppThemeProvider } from '@hpapp/contexts/settings/theme';
 import { initURICache } from '@hpapp/features/common/hooks/uricache';
 import Initialize from '@hpapp/features/root/protected/Initializer';
 import LoadError from '@hpapp/features/root/protected/LoadError';
 import RootWrapper from '@hpapp/features/root/protected/RootWrapper';
+import { ServiceRootProvider } from '@hpapp/features/root/protected/context';
 import { Screen, ScreenParams, createStackNavigator } from '@hpapp/features/root/protected/stack';
+import { AppThemeProvider } from '@hpapp/features/settings/context/theme';
 import { useNavigationContainerRef } from '@react-navigation/native';
 import { Text } from '@rneui/base';
 import React from 'react';
