+++ expo/features/root/protected/RootWrapper.tsx
@@ -1,4 +1,4 @@
-import { useColor } from '@hpapp/contexts/settings/theme';
+import { useColor } from '@hpapp/features/settings/context/theme';
 import { View } from 'react-native';
 
 export default function RootWrapper({ children }: { children: React.ReactElement }) {
