+++ expo/features/settings/devonly/DevOnlySettingsScreen.tsx
@@ -1,6 +1,6 @@
-import useLocalUserConfig from '@hpapp/contexts/settings/useLocalUserConfig';
 import { useCurrentUser } from '@hpapp/features/auth';
 import { defineScreen } from '@hpapp/features/root/protected/stack';
+import useLocalUserConfig from '@hpapp/features/settings/context/useLocalUserConfig';
 import DevOnlySettingsListItem from '@hpapp/features/settings/devonly/DevOnlySettingsListItem';
 import { Divider } from '@rneui/themed';
