+++ expo/features/settings/theme/ThemeSettingsScreen.tsx
@@ -1,8 +1,8 @@
-import { ColorScheme, useColor } from '@hpapp/contexts/settings/theme';
 import Text from '@hpapp/features/common/components/Text';
 import NavigationListItem from '@hpapp/features/common/components/list/NavigationListItem';
 import { Spacing } from '@hpapp/features/common/constants';
 import { defineScreen, useScreenTitle } from '@hpapp/features/root/protected/stack';
+import { ColorScheme, useColor } from '@hpapp/features/settings/context/theme';
 import ThemeColorSelectorScreen from '@hpapp/features/settings/theme/ThemeColorSelectorScreen';
 import { t } from '@hpapp/system/i18n';
 import { View, StyleSheet } from 'react-native';
